# openapi.yaml
openapi: 3.0.3
info:
  title: NOC Chamados API
  description: |
    API REST para sistema de gestão de chamados de NOC (Network Operations Center).
    
    ## Autenticação
    A API usa autenticação JWT via HTTPOnly Cookie. Faça login primeiro usando `/auth/login`.
    
    ## Rate Limiting
    - Login: 5 tentativas por 15 minutos (por IP)
    
    ## Recursos
    - [Documentação Interativa](/api-docs)
    - [Guia Swagger](./SWAGGER_GUIDE.md)
    - [Coleção Postman](./NOC_Chamados.postman_collection.json)
    
  version: 1.2.3
  contact:
    name: Suporte NOC Chamados
  license:
    name: Proprietário
servers:
  - url: https://chamados.xmov.com.br/api
    description: Produção
  - url: http://localhost:3000/api
    description: Desenvolvimento Local
...
paths:
...
  /transportes:
    get:
      tags:
        - Transportes
      summary: Listar Transportes
      operationId: listarTransportes
      security:
        - cookieAuth: []
      parameters:
        - name: ativo
          in: query
          schema:
            type: boolean
        - name: includeLinks
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Lista de transportes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transporte'

    post:
      tags:
        - Transportes
      summary: Criar Transporte
      operationId: criarTransporte
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CriarTransporte'
      responses:
        '201':
          description: Transporte criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transporte'
  /transportes/{id}:
    get:
      tags:
        - Transportes
      summary: Obter Transporte
      operationId: obterTransporte
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Detalhes do transporte
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transporte'

    put:
      tags:
        - Transportes
      summary: Atualizar Transporte
      operationId: atualizarTransporte
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CriarTransporte'
      responses:
        '200':
          description: Transporte atualizado

    delete:
      tags:
        - Transportes
      summary: Deletar Transporte
      operationId: deletarTransporte
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Transporte deletado
...
  /usuarios:
    get:
      tags:
        - Usuários
      summary: Listar Usuários
      description: Lista usuários (apenas supervisores)
      operationId: listarUsuarios
      security:
        - cookieAuth: []
      parameters:
        - name: busca
          in: query
          schema:
            type: string
        - name: inativos
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Lista de usuários
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Usuario'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags:
        - Usuários
      summary: Criar Usuário
      description: Cria um novo usuário (apenas supervisores)
      operationId: criarUsuario
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CriarUsuario'
      responses:
        '201':
          description: Usuário criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        '403':
          $ref: '#/components/responses/Forbidden'
  /usuarios/{id}:
    get:
      tags:
        - Usuários
      summary: Obter Usuário
      operationId: obterUsuario
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Detalhes do usuário
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'

    put:
      tags:
        - Usuários
      summary: Atualizar Usuário
      operationId: atualizarUsuario
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AtualizarUsuario'
      responses:
        '200':
          description: Usuário atualizado

    delete:
      tags:
        - Usuários
      summary: Deletar Usuário
      description: Deleta um usuário (hard delete se sem chamados, soft delete se tiver)
      operationId: deletarUsuario
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Usuário deletado
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: JWT token armazenado em HTTPOnly cookie
  responses:
...
    Forbidden:
      description: Sem permissão
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Acesso negado
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: Mensagem de erro
...
    Transporte:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nome:
          type: string
          example: TRANSPORTE-001
        fornecedor:
          type: string
          nullable: true
          example: Fornecedor A
        origem:
          type: string
          nullable: true
          example: São Paulo
        destino:
          type: string
          nullable: true
          example: Rio de Janeiro
        ativo:
          type: boolean
          example: true
        observacoes:
          type: string
          nullable: true
        _count:
          type: object
          properties:
            links:
              type: integer
            chamados:
              type: integer
    CriarTransporte:
      type: object
      required:
        - nome
      properties:
        nome:
          type: string
          example: TRANSPORTE-001
        fornecedor:
          type: string
          example: Fornecedor A
        origem:
          type: string
          example: São Paulo
        destino:
          type: string
          example: Rio de Janeiro
        observacoes:
          type: string
...
    Usuario:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nome:
          type: string
          example: João Silva
        email:
          type: string
          format: email
          example: joao@xmov.com.br
        perfil:
          type: string
          enum: [operador, supervisor]
          example: operador
        ativo:
          type: boolean
          example: true
        criadoEm:
          type: string
          format: date-time
        _count:
          type: object
          properties:
            chamadosAbertos:
              type: integer
            chamadosFechados:
              type: integer
...
    CriarUsuario:
      type: object
      required:
        - nome
        - email
        - senha
        - perfil
      properties:
        nome:
          type: string
          example: João Silva
        email:
          type: string
          format: email
          example: joao@xmov.com.br
        senha:
          type: string
          format: password
          minLength: 8
          pattern: '^(?=.*[a-z])(?=.*[A-Z])(?=.*\d).+$'
          example: Senha@123
        perfil:
          type: string
          enum: [operador, supervisor]
          example: operador
    AtualizarUsuario:
      type: object
      properties:
        nome:
          type: string
          example: João Silva
        email:
          type: string
          format: email
          example: joao@xmov.com.br
        perfil:
          type: string
          enum: [operador, supervisor]
...
security:
  - cookieAuth: []